backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
Pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
torch==2.1.2
torchvision==0.16.2
ultralytics==8.3.170
opencv-python-headless==4.9.0.80
pandas==2.1.4
joblib==1.3.2
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
